refactor: better naming
Rename the install method to fetch. Move a method out of installMultiple
and call it install instead.
