+++ src/api/install.ts
@@ -16,7 +16,7 @@ import getSaveType from '../getSaveType'
 import {sync as runScriptSync} from '../runScript'
 import postInstall from '../install/postInstall'
 import extendOptions from './extendOptions'
-import {InstalledPackage} from '../install'
+import {InstalledPackage} from '../install/fetch'
 import pnpmPkgJson from '../pnpmPkgJson'
 import lock from './lock'
 import {save as saveStore, Store} from '../fs/storeController'
