+++ src/install/installMultiple.ts
@@ -1,5 +1,5 @@
 import path = require('path')
-import install, {InstalledPackage, InstallationOptions} from './'
+import fetch, {InstalledPackage, InstallationOptions} from './fetch'
 import {InstallContext, InstalledPackages} from '../api/install'
 import {Dependencies} from '../types'
 import linkBins from './linkBins'
@@ -25,52 +25,7 @@ export default async function installMultiple (ctx: InstallContext, pkgsMap: Dep
   const installedPkgs: InstalledPackage[] = <InstalledPackage[]>(
     await Promise.all(pkgs.map(async function (pkgRawSpec: string) {
       try {
-        const dependency = await install(ctx.fetches, pkgRawSpec, modules, options)
-
-        addInstalledPkg(ctx.installs, dependency)
-
-        ctx.store.packages[options.dependent] = ctx.store.packages[options.dependent] || {}
-        ctx.store.packages[options.dependent].dependencies = ctx.store.packages[options.dependent].dependencies || {}
-
-        // NOTE: the current install implementation
-        // does not return enough info for packages that were already installed
-        if (!dependency.fromCache) {
-          ctx.store.packages[options.dependent].dependencies[dependency.pkg.name] = dependency.id
-
-          ctx.store.packages[dependency.id] = ctx.store.packages[dependency.id] || {}
-          ctx.store.packages[dependency.id].dependents = ctx.store.packages[dependency.id].dependents || []
-          if (ctx.store.packages[dependency.id].dependents.indexOf(options.dependent) === -1) {
-            ctx.store.packages[dependency.id].dependents.push(options.dependent)
-          }
-
-          if (dependency.justFetched || dependency.firstFetch && dependency.keypath.length <= options.depth) {
-            const nextInstallOpts = Object.assign({}, options, {
-                keypath: dependency.keypath.concat([ dependency.id ]),
-                dependent: dependency.id,
-                optional: dependency.optional,
-                root: dependency.srcPath,
-              })
-            dependency.dependencies = (await installMultiple(ctx,
-                dependency.pkg.dependencies || {},
-                path.join(dependency.path, 'node_modules'),
-                nextInstallOpts)
-            ).concat(
-              await installMultiple(ctx,
-                dependency.pkg.optionalDependencies || {},
-                path.join(dependency.path, 'node_modules'),
-                Object.assign({}, nextInstallOpts, {
-                  optional: true
-                }))
-            )
-            ctx.piq = ctx.piq || []
-            ctx.piq.push({
-              path: dependency.path,
-              pkgId: dependency.id
-            })
-          }
-        }
-
-        return dependency
+        return install(pkgRawSpec, modules, ctx, options)
       } catch (err) {
         if (options.optional) {
           console.log(`Skipping failed optional dependency ${pkgRawSpec}:`)
@@ -86,6 +41,55 @@ export default async function installMultiple (ctx: InstallContext, pkgsMap: Dep
   await linkBins(modules)
 
   return installedPkgs
+}
+
+async function install (pkgRawSpec: string, modules: string, ctx: InstallContext, options: MultipleInstallationOptions) {
+  const dependency = await fetch(ctx.fetches, pkgRawSpec, modules, options)
+
+  addInstalledPkg(ctx.installs, dependency)
+
+  ctx.store.packages[options.dependent] = ctx.store.packages[options.dependent] || {}
+  ctx.store.packages[options.dependent].dependencies = ctx.store.packages[options.dependent].dependencies || {}
+
+  // NOTE: the current install implementation
+  // does not return enough info for packages that were already installed
+  if (!dependency.fromCache) {
+    ctx.store.packages[options.dependent].dependencies[dependency.pkg.name] = dependency.id
+
+    ctx.store.packages[dependency.id] = ctx.store.packages[dependency.id] || {}
+    ctx.store.packages[dependency.id].dependents = ctx.store.packages[dependency.id].dependents || []
+    if (ctx.store.packages[dependency.id].dependents.indexOf(options.dependent) === -1) {
+      ctx.store.packages[dependency.id].dependents.push(options.dependent)
+    }
+
+    if (dependency.justFetched || dependency.firstFetch && dependency.keypath.length <= options.depth) {
+      const nextInstallOpts = Object.assign({}, options, {
+          keypath: dependency.keypath.concat([ dependency.id ]),
+          dependent: dependency.id,
+          optional: dependency.optional,
+          root: dependency.srcPath,
+        })
+      dependency.dependencies = (await installMultiple(ctx,
+          dependency.pkg.dependencies || {},
+          path.join(dependency.path, 'node_modules'),
+          nextInstallOpts)
+      ).concat(
+        await installMultiple(ctx,
+          dependency.pkg.optionalDependencies || {},
+          path.join(dependency.path, 'node_modules'),
+          Object.assign({}, nextInstallOpts, {
+            optional: true
+          }))
+      )
+      ctx.piq = ctx.piq || []
+      ctx.piq.push({
+        path: dependency.path,
+        pkgId: dependency.id
+      })
+    }
+  }
+
+  return dependency
 }
 
 function addInstalledPkg (installs: InstalledPackages, newPkg: InstalledPackage) {
