+++ src/resolve/local.ts
@@ -5,8 +5,7 @@ import createPkgId, {delimiter} from './createPkgId'
 import getTarballName from './getTarballName'
 import requireJson from '../fs/requireJson'
 import mkdirp from '../fs/mkdirp'
-import {PackageSpec} from '../install'
-import {ResolveOptions, ResolveResult} from '.'
+import {PackageSpec, ResolveOptions, ResolveResult} from '.'
 import {createLocalTarballFetcher, fetchFromLocalTarball} from './fetch'
 import fs = require('mz/fs')
 import linkDir from 'link-dir'
