+++ tsconfig.json
@@ -52,8 +52,8 @@
     "src/fs/unsymlink.ts",
     "src/getSaveType.ts",
     "src/index.ts",
+    "src/install/fetch.ts",
     "src/install/flattenDependencies.ts",
-    "src/install/index.ts",
     "src/install/installMultiple.ts",
     "src/install/isAvailable.ts",
     "src/install/linkBins.ts",
